// components/admin/directory/adminDictionaryKPI/adminDictionaryKPIEditor/adminDictionaryKPIEditorStyle.scss.d.ts
export type Styles = {
  'checkBoxSelectedAllPositions': string;
  'dynamicFormulaEditorsContainer': string;
  'errorTextDropDownSelect': string;
  'inputTextWrapper': string;
  'popupBodySize': string;
  'selectedAllPositionsHelpText': string;
};

export type ClassNames = keyof Styles;

declare const styles: Styles;

export default styles;